+++ backend/connector-integration/src/connectors/adyen/test.rs
@@ -62,6 +62,7 @@ mod tests {
                         adyen: ConnectorParams {
                             base_url: "https://checkout-test.adyen.com/".to_string(),
                             dispute_base_url: Some("https://ca-test.adyen.com/ca/services/DisputeService/v30/defendDispute".to_string()),
+                            ..Default::default()
                         },
                         ..Default::default()
                     },
@@ -242,6 +243,7 @@ mod tests {
                         adyen: ConnectorParams {
                             base_url: "https://checkout-test.adyen.com/".to_string(),
                             dispute_base_url: Some("https://ca-test.adyen.com/ca/services/DisputeService/v30/defendDispute".to_string()),
+                            ..Default::default()
                         },
                         ..Default::default()
                     },