+++ backend/domain_types/src/connector_flow.rs
@@ -55,6 +55,9 @@ pub struct Authenticate;
 #[derive(Debug, Clone)]
 pub struct PostAuthenticate;
 
+#[derive(Debug, Clone)]
+pub struct VoidPC;
+
 #[derive(strum::Display)]
 #[strum(serialize_all = "snake_case")]
 pub enum FlowName {
@@ -63,6 +66,7 @@ pub enum FlowName {
     Rsync,
     Psync,
     Void,
+    VoidPc,
     SetupMandate,
     RepeatPayment,
     Capture,